Fix check existence of csv generated file

+++ labonneboite/scripts/impact_retour_emploi/join_activity_logs_dpae.py
@@ -159,11 +159,12 @@ def join_dpae_activity_logs(df_activity):
 
         timestr = time.strftime("%Y-%m-%d")
         path_to_csv = f"{dpae_folder_path}act_dpae-{timestr}.csv"
-        #file_exists = os.path.isfile(path_to_csv)
+        file_exists = os.path.isfile(path_to_csv)
 
         # We want to rewrite the CSV file after each new execution of DPAE extraction
         if i == 0:
-            os.remove(path_to_csv)
+            if file_exists:
+                os.remove(path_to_csv)
             df_dpae_act.to_csv(path_to_csv, encoding='utf-8', sep='|')
         else:
             with open(path_to_csv, 'a') as f:
